test: Huge testing overhaul
From now on I'm trying to apply the following test conventions: 1. All tests have a testDefaultBehavior test to verify that the function works properly. 2. Fuzzing should always be used when possible. 3. All the states changed by test should be checked. 4. Testing the constructor default behavior comes down to testing that the setup function changed the state of the contrat as expected